--- a/Cargo.toml
+++ b/Cargo.toml
@@ -36,8 +36,7 @@ path = "examples/multiple_input.rs"
 
 [dependencies]
 bevy = { version = "0.13", default-features = false, features = [] }
-# trie-rs = { version = "0.3" }
-trie-rs = { path = "../trie-rs" }
+trie-rs = { version = "0.4" }
 keyseq = { version = "0.2.3", features = [ "bevy" ] }
 
 [dev-dependencies]