--- a/src/lib.rs
+++ b/src/lib.rs
@@ -27,11 +27,10 @@ pub use keyseq::{
 
 /// Convenient glob import
 pub mod prelude {
-    pub use std::time::Duration;
+    pub use super::input_sequence::{ButtonSequence, InputSequence, KeySequence};
+    pub use super::{action, keyseq, InputSequencePlugin, Modifiers, TimeLimit};
 
-    pub use crate::input_sequence::{ButtonSequence, InputSequence, KeySequence};
-
-    pub use super::{action, InputSequencePlugin, keyseq, Modifiers, TimeLimit};
-    pub use super::cond_system::IntoCondSystem;
     pub use super::chord::KeyChord;
+    pub use super::cond_system::IntoCondSystem;
+    pub use std::time::Duration;
 }