Hey Elm,

This is a performance enhancement and maybe a small behavioral fix.

I've done some work on [trie-rs](https://github.com/laysakura/trie-rs/blob/4a334a10725a128ae27edb9360842d9593580433/CHANGELOG.md#v040) which has finally come to fruition. A trie is the right data structure for what we're doing. It's look ups are _O(m log n)_ where _m_ is the key chord count and _n_ is the max number of leaves to a node in the trie. This isn't bad, but we do it once per key and we also check all our postfixes `buffer[i..]` for _i \in [0, m]_, which I realize now may actually be slightly incorrect behavior I'll explain below. So our actual big O probably more like _O(m^2 log n)_.

Now the silly thing is we should be able to do everything we need for each key press in _O(log n)_ if we just keep track of where we are in the trie; we just need to look at our leaves. And that's essentially what this patch does. I explain more about what's under the covers in the [trie incremental search documentation](https://docs.rs/trie-rs/0.4.0/trie_rs/inc_search/index.html).

## Postfix Bug

If you recognize key sequences "W A S D" and "A S" and some one types "W A S P", you probably don't want to recognize any key sequences. Well, our current code would recognize that as a match for the "A S" sequence. This patch once it detects that there is no match for the sequence will reset and start looking for sequences starting with "P".

## Allocating Less Too

Previously we allocated on each key press in `consume_input` using a `Vec`. We're now using an iterator so no allocation there.

Since this doesn't touch the API surface at all, you can probably just bump it to 0.4.1 I would think. But that's at your discretion of course.

# Future Idea

One idea to consider for the future is making the key sequence that was pressed available somehow. Maybe something simple like this:

```rust
#[derive(Resource)]
struct LastKeySequence(Option<Vec<KeyChord>>)
```

But maybe that's too simple. What'd be best is if we could optionally pass the key sequence to the system we're invoking. But I don't really know how to do that.