@not-elm 2024-05-01
I have created a simple [benchmark](https://github.com/not-elm/bevy-input-sequence/tree/benchmark).

I checked and performance was improved, so this fix seems not a problem.
```
before(master)     :    [826.48 µs 841.84 µs 856.72 µs]
after(this branch) :    [783.06 µs 798.71 µs 819.33 µs]
```

Regarding `Future Idea`: what is the use of `LastKeySequence`, for example?