@shanecelis 2024-05-01
That's really cool benchmark code. You're a wizard with these test harnesses, @not-elm. I wouldn't have known how to start that at all. 

Re: LastKeySequence. It's admittedly for bevy_minibuffer. Sometimes commands want to know what key sequence fired them as there can be multiple bindings for one command. Maybe it'd be better to have some kind of "hook" that can run so the user can create things like that for themselves. Some kind of `FnMut(KeyChord, trie_rs::inc_search::Answer)` maybe.