@not-elm 2024-05-03
Hmm, sorry, but I honestly can't yet imagine how `Hook` would work.
I have created a dedicated issue [here](https://github.com/not-elm/bevy-input-sequence/issues/8#7). Can you give us a simple sample code there?

Also, I will merge this PR itself.